+++ src/connections/api.d.ts
@@ -1,4 +1,4 @@
-/* eslint-disable node/no-missing-import */
+ 
 import { Response } from '../response';
 import { Connection, UserConnectionCredentials, ConnectionCollection } from './dtos';