+++ src/connections/dtos.d.ts
@@ -1,8 +1,4 @@
-/* eslint-disable no-shadow */
-/* eslint-disable no-redeclare */
-/* eslint-disable @typescript-eslint/no-namespace */
-
-// eslint-disable-next-line node/no-missing-import
+ 
 import { IPaginated } from '../response';
 
 /**
@@ -73,10 +69,10 @@ export interface Connection {
     defaultConnectionProperties?: ConnectionDefaultConnectionProperties;
     /**
      *
-     * @type {Date}
+     * @type {string}
      * @memberof Connection
      */
-    createdTime?: Date;
+    createdTime?: string;
 }
 
 /**