+++ src/tenants/dtos.d.ts
@@ -1,8 +1,4 @@
-/* eslint-disable no-shadow */
-/* eslint-disable no-redeclare */
-/* eslint-disable @typescript-eslint/no-namespace */
-
-// eslint-disable-next-line node/no-missing-import
+ 
 import { IPaginated } from '../response';
 
 /**
@@ -37,10 +33,10 @@ export interface Tenant {
   connection?: TenantConnection;
   /**
    *
-   * @type {Date}
+   * @type {string}
    * @memberof Tenant
    */
-  createdTime?: Date;
+  createdTime?: string;
 }
 /**
 * A collection of tenants.