+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change log
 This is the changelog for [Authress SDK](readme.md).
 
+## 3.1 ##
+* Require minimum Node version to be 20.
+* Add Support for KMS Service Client BYOK usage.
+
 ## 3.0 ##
 * Require minimum Node version to be 18.
 * [Breaking] UserId is now required in all `userPermissions` apis. This improves **Security By Default** requiring explicit check on who the user is.